Move GW2Builds to its own file with release notes documentation, moved ChestIDs to SpeciesIDs
