+++ GW2EIEvtcParser/EIData/Buffs/Buff.cs
@@ -2,6 +2,7 @@
 using GW2EIEvtcParser.EIData.BuffSourceFinders;
 using GW2EIEvtcParser.Interfaces;
 using GW2EIEvtcParser.ParsedData;
+using GW2EIEvtcParser.ParserHelpers;
 using static GW2EIEvtcParser.ArcDPSEnums;
 using static GW2EIEvtcParser.ParserHelper;
