+++ GW2EIEvtcParser/EIData/Buffs/BuffInfoSolver.cs
@@ -1,5 +1,6 @@
 ﻿using GW2EIEvtcParser.Interfaces;
 using GW2EIEvtcParser.ParsedData;
+using GW2EIEvtcParser.ParserHelpers;
 using static GW2EIEvtcParser.ArcDPSEnums;
 using static GW2EIEvtcParser.SkillIDs;
