+++ GW2EIEvtcParser/EIData/DamageModifiers/DamageModifierDescriptors/DamageModifierDescriptor.cs
@@ -1,6 +1,7 @@
 ﻿using GW2EIEvtcParser.EncounterLogic;
 using GW2EIEvtcParser.Interfaces;
 using GW2EIEvtcParser.ParsedData;
+using GW2EIEvtcParser.ParserHelpers;
 using static GW2EIEvtcParser.ArcDPSEnums;
 using static GW2EIEvtcParser.EIData.DamageModifiersUtils;
 using static GW2EIEvtcParser.ParserHelper;
