+++ GW2EIEvtcParser/EncounterLogic/EncounterLogicTimeUtils.cs
@@ -2,6 +2,7 @@
 using GW2EIEvtcParser.Exceptions;
 using GW2EIEvtcParser.ParsedData;
 using static GW2EIEvtcParser.ArcDPSEnums;
+using static GW2EIEvtcParser.SpeciesIDs;
 using static GW2EIEvtcParser.EncounterLogic.EncounterLogicUtils;
 
 namespace GW2EIEvtcParser.EncounterLogic;
