+++ GW2EIEvtcParser/EncounterLogic/Fractals/SunquaPeak/SunquaPeak.cs
@@ -1,5 +1,6 @@
 ﻿using GW2EIEvtcParser.Extensions;
 using GW2EIEvtcParser.ParsedData;
+using GW2EIEvtcParser.ParserHelpers;
 using static GW2EIEvtcParser.ArcDPSEnums;
 using static GW2EIEvtcParser.EncounterLogic.EncounterCategory;
